Organize _kaleido_tab exports.
